Removed CLion warnings

+++ utility/mapcode.cpp
@@ -49,37 +49,35 @@
 #define my_isnan(x) (false)
 #define my_round(x) ((int) (floor((x) + 0.5)))
 
-static int          selfCheckEnabled = 0;
+static int selfCheckEnabled = 0;
 
-static const int    NORMAL_ERROR     = 1;
-static const int    INTERNAL_ERROR   = 2;
+static const int NORMAL_ERROR = 1;
+static const int INTERNAL_ERROR = 2;
 
 /**
  * Some global constants to be used.
  */
-static const double PI              = 3.14159265358979323846;
-static const int    SHOW_PROGRESS   = 125;
-static const double DELTA           = 0.001;
-
+static const double PI = 3.14159265358979323846;
+static const int SHOW_PROGRESS = 125;
+static const double DELTA = 0.001;
 
 
 /**
  * These statistics are stored globally so they can be updated easily by the
  * generateAndOutputMapcodes() method.
  */
-static int      totalNrOfPoints         = 0;
-static int      totalNrOfResults        = 0;
-static int      largestNrOfResults      = 0;
-static double   latLargestNrOfResults   = 0.0;
-static double   lonLargestNrOfResults   = 0.0;
-
+static int totalNrOfPoints = 0;
+static int totalNrOfResults = 0;
+static int largestNrOfResults = 0;
+static double latLargestNrOfResults = 0.0;
+static double lonLargestNrOfResults = 0.0;
 
 
 /**
  * The usage() method explains how this application can be used. It is called
  * whenever a incorrect amount or combination of parameters is entered.
  */
-static void usage(const char* appName) {
+static void usage(const char *appName) {
     printf("MAPCODE (version %s)\n", mapcode_cversion);
     printf("Copyright (C) 2014-2015 Stichting Mapcode Foundation\n");
     printf("\n");
@@ -140,23 +138,20 @@ static void usage(const char* appName) {
 }
 
 
-
 /**
  * The method radToDeg() converts radians to degrees.
  */
-static double radToDeg(double rad){
+static double radToDeg(double rad) {
     return (rad / PI) * 180.0;
 }
-
 
 
 /**
  * The method degToRad() converts degrees to radians.
  */
-static double degToRad(double deg){
+static double degToRad(double deg) {
     return (deg / 180.0) * PI;
 }
-
 
 
 /**
@@ -164,7 +159,7 @@ static double degToRad(double deg){
  * (x, y, z) point on a sphere with a radius of 1.
  */
 static void unitToLatLonDeg(
-    const double unit1, const double unit2, double* latDeg, double* lonDeg) {
+        const double unit1, const double unit2, double *latDeg, double *lonDeg) {
 
     // Calculate uniformly distributed 3D point on sphere (radius = 1.0):
     // http://mathproofs.blogspot.co.il/2005/04/uniform-random-distribution-on-sphere.html
@@ -185,12 +180,11 @@ static void unitToLatLonDeg(
 }
 
 
-
 /**
  * The method convertLatLonToXYZ() convertes a lat/lon pair to a (x, y, z) coordinate
  * on a sphere with radius 1.
  */
-static void convertLatLonToXYZ(double latDeg, double lonDeg, double* x, double* y, double* z) {
+static void convertLatLonToXYZ(double latDeg, double lonDeg, double *x, double *y, double *z) {
     double latRad = degToRad(latDeg);
     double lonRad = degToRad(lonDeg);
     *x = cos(latRad) * cos(lonRad);
@@ -199,20 +193,20 @@ static void convertLatLonToXYZ(double latDeg, double lonDeg, double* x, double* 
 }
 
 
-
 /**
  * This methods provides a self check for encoding lat/lon to Mapcode.
  */
-static void selfCheckLatLonToMapcode(const double lat, double lon, const char* territory, const char* mapcode, int extraDigits) {
+static void selfCheckLatLonToMapcode(const double lat, double lon, const char *territory, const char *mapcode,
+                                     int extraDigits) {
     int context = convertTerritoryIsoNameToCode(territory, 0);
-    char* results[2 * MAX_NR_OF_MAPCODE_RESULTS];
+    char *results[2 * MAX_NR_OF_MAPCODE_RESULTS];
     const double limitLat = (lat < -90.0) ? -90.0 : ((lat > 90.0) ? 90.0 : lat);
     const double limitLon = (lon < -180.0) ? -180.0 : ((lon > 180.0) ? 180.0 : lon);
     const int nrResults = encodeLatLonToMapcodes_Deprecated(results, limitLat, limitLon, context, extraDigits);
     if (nrResults <= 0) {
         fprintf(stderr, "error: encoding lat/lon to mapcode failure; "
-            "cannot encode lat=%.12g, lon=%.12g (default territory=%s)\n",
-            lat, lon, territory);
+                        "cannot encode lat=%.12g, lon=%.12g (default territory=%s)\n",
+                lat, lon, territory);
         if (selfCheckEnabled) {
             exit(INTERNAL_ERROR);
         }
@@ -224,22 +218,22 @@ static void selfCheckLatLonToMapcode(const double lat, double lon, const char* t
         /* Check if the territory and code were found in results. Note that the territory
          * may be a minimal code, like IN (which may indicate US-IN or RU-IN).
          */
-        const char* foundMapcode = results[(i * 2)];
-        const char* foundTerritory = results[(i * 2) + 1];
-        char* foundTerritoryMin = strstr(foundTerritory, "-");
+        const char *foundMapcode = results[(i * 2)];
+        const char *foundTerritory = results[(i * 2) + 1];
+        char *foundTerritoryMin = strstr(foundTerritory, "-");
         if (foundTerritoryMin && (strlen(foundTerritoryMin) > 0)) {
             ++foundTerritoryMin;
         }
 
         found = (((strcmp(territory, foundTerritory) == 0) ||
-                    (strcmp(territory, foundTerritoryMin) == 0)) &&
-                (strcmp(mapcode, foundMapcode) == 0));
+                  (strcmp(territory, foundTerritoryMin) == 0)) &&
+                 (strcmp(mapcode, foundMapcode) == 0));
     }
     if (!found) {
         fprintf(stderr, "error: encoding lat/lon to mapcode failure; "
-            "mapcode '%s %s' decodes to lat=%.12g(%.12g), lon=%.12g(%.12g), "
-            "which does not encode back to '%s %s'\n",
-            territory, mapcode, lat, limitLat, lon, limitLon, territory, mapcode);
+                        "mapcode '%s %s' decodes to lat=%.12g(%.12g), lon=%.12g(%.12g), "
+                        "which does not encode back to '%s %s'\n",
+                territory, mapcode, lat, limitLat, lon, limitLon, territory, mapcode);
         if (selfCheckEnabled) {
             exit(INTERNAL_ERROR);
         }
@@ -248,34 +242,33 @@ static void selfCheckLatLonToMapcode(const double lat, double lon, const char* t
 }
 
 
-
 /**
  * This method provides a self-check for decoding a Mapcode to lat/lon.
  */
-static void selfCheckMapcodeToLatLon(const char* territory, const char* mapcode,
-    const double lat, const double lon) {
+static void selfCheckMapcodeToLatLon(const char *territory, const char *mapcode,
+                                     const double lat, const double lon) {
     double foundLat;
     double foundLon;
     int foundContext = convertTerritoryIsoNameToCode(territory, 0);
     int err = decodeMapcodeToLatLon(&foundLat, &foundLon, mapcode, foundContext);
     if (err != 0) {
         fprintf(stderr, "error: decoding mapcode to lat/lon failure; "
-            "cannot decode '%s %s')\n", territory, mapcode);
+                "cannot decode '%s %s')\n", territory, mapcode);
         if (selfCheckEnabled) {
             exit(INTERNAL_ERROR);
         }
         return;
     }
     double deltaLat = ((foundLat - lat) >= 0.0 ? (foundLat - lat) : -(foundLat - lat));
-    double deltaLon = ((foundLon - lon) >- 0.0 ? (foundLon - lon) : -(foundLon - lon));
+    double deltaLon = ((foundLon - lon) > -0.0 ? (foundLon - lon) : -(foundLon - lon));
     if (deltaLon > 180.0) {
         deltaLon = 360.0 - deltaLon;
     }
     if ((deltaLat > DELTA) || (deltaLon > DELTA)) {
         fprintf(stderr, "error: decoding mapcode to lat/lon failure; "
-            "lat=%.12g, lon=%.12g produces mapcode %s %s, "
-            "which decodes to lat=%.12g (delta=%.12g), lon=%.12g (delta=%.12g)\n",
-            lat, lon, territory, mapcode, foundLat, deltaLat, foundLon, deltaLon);
+                        "lat=%.12g, lon=%.12g produces mapcode %s %s, "
+                        "which decodes to lat=%.12g (delta=%.12g), lon=%.12g (delta=%.12g)\n",
+                lat, lon, territory, mapcode, foundLat, deltaLat, foundLon, deltaLon);
         if (selfCheckEnabled) {
             exit(INTERNAL_ERROR);
         }
@@ -285,7 +278,7 @@ static void selfCheckMapcodeToLatLon(const char* territory, const char* mapcode,
 
 static void generateAndOutputMapcodes(double lat, double lon, int iShowError, int extraDigits, int useXYZ) {
 
-    char* results[2 * MAX_NR_OF_MAPCODE_RESULTS];
+    char *results[2 * MAX_NR_OF_MAPCODE_RESULTS];
     int context = 0;
 
     while (lon > 180.0) {
@@ -330,8 +323,8 @@ static void generateAndOutputMapcodes(double lat, double lon, int iShowError, in
         printf("%d %.12g %.12g\n", nrResults, lat, lon);
     }
     for (int j = 0; j < nrResults; ++j) {
-        const char* foundMapcode = results[(j * 2)];
-        const char* foundTerritory = results[(j * 2) + 1];
+        const char *foundMapcode = results[(j * 2)];
+        const char *foundTerritory = results[(j * 2) + 1];
 
         // Output result line.
         printf("%s %s\n", foundTerritory, foundMapcode);
@@ -355,7 +348,6 @@ static void generateAndOutputMapcodes(double lat, double lon, int iShowError, in
 }
 
 
-
 /**
  * This method resets the statistics counters.
  */
@@ -367,7 +359,6 @@ static void resetStatistics(int nrOfPoints) {
 }
 
 
-
 /**
  * This method outputs the statistics.
  */
@@ -376,11 +367,10 @@ static void outputStatistics() {
     fprintf(stderr, "Total number of 3D points generated     = %d\n", totalNrOfPoints);
     fprintf(stderr, "Total number of mapcodes generated      = %d\n", totalNrOfResults);
     fprintf(stderr, "Average number of mapcodes per 3D point = %.12g\n",
-        ((float) totalNrOfResults) / ((float) totalNrOfPoints));
+            ((float) totalNrOfResults) / ((float) totalNrOfPoints));
     fprintf(stderr, "Largest number of results for 1 mapcode = %d at (%.12g, %.12g)\n",
-        largestNrOfResults, latLargestNrOfResults, lonLargestNrOfResults);
+            largestNrOfResults, latLargestNrOfResults, lonLargestNrOfResults);
 }
-
 
 
 /**
@@ -388,18 +378,16 @@ static void outputStatistics() {
  */
 static void showProgress(int i) {
     fprintf(stderr, "[%d%%] Processed %d of %d regions (generated %d mapcodes)...\r",
-        (int) ((((float) i / ((float) totalNrOfPoints)) * 100.0) + 0.5),
-        i, totalNrOfPoints, totalNrOfResults);
+            (int) ((((float) i / ((float) totalNrOfPoints)) * 100.0) + 0.5),
+            i, totalNrOfPoints, totalNrOfResults);
 }
-
 
 
 /**
  * This is the main() method which is called from the command-line.
  * Return code 0 means success. Any other values means some sort of error occurred.
  */
-int main(const int argc, const char** argv)
-{
+int main(const int argc, const char **argv) {
     // Assume no extra digits (unless overridden later.
     int extraDigits = 0;
 
@@ -407,7 +395,7 @@ int main(const int argc, const char** argv)
     int useXYZ = 0;
 
     // Provide usage message if no arguments specified.
-    const char* appName = argv[0];
+    const char *appName = argv[0];
     selfCheckEnabled = (strstr(appName, "debug") != 0);
     if (selfCheckEnabled) {
         fprintf(stderr, "(debug mode: self checking enabled)\n");
@@ -418,7 +406,7 @@ int main(const int argc, const char** argv)
     }
 
     // First argument: command.
-    const char* cmd = argv[1];
+    const char *cmd = argv[1];
     if ((strcmp(cmd, "-d") == 0) || (strcmp(cmd, "--decode") == 0)) {
 
         // ------------------------------------------------------------------
@@ -430,7 +418,7 @@ int main(const int argc, const char** argv)
             return NORMAL_ERROR;
         }
 
-        const char* defaultTerritory = argv[2];
+        const char *defaultTerritory = argv[2];
         double lat;
         double lon;
 
@@ -441,7 +429,7 @@ int main(const int argc, const char** argv)
         for (int i = 3; i < argc; ++i) {
 
             // Decode the Mapcode to a lat/lon.
-            const char* mapcode = argv[i];
+            const char *mapcode = argv[i];
             int err = decodeMapcodeToLatLon(&lat, &lon, mapcode, context);
             if (err != 0) {
                 fprintf(stderr, "error: cannot decode '%s %s'\n", defaultTerritory, mapcode);
@@ -453,7 +441,7 @@ int main(const int argc, const char** argv)
 
             // Self-checking code to see if encoder produces this Mapcode for the lat/lon.
             if (selfCheckEnabled) {
-                const char* suffix = strstr(mapcode, "-");
+                const char *suffix = strstr(mapcode, "-");
                 extraDigits = 0;
                 if (suffix != 0) {
                     extraDigits = (int) (strlen(suffix) - 1);
@@ -463,15 +451,15 @@ int main(const int argc, const char** argv)
         }
     }
     else if ((strcmp(cmd, "-e") == 0) || (strcmp(cmd, "-e0") == 0) ||
-        (strcmp(cmd, "-e1") == 0) || (strcmp(cmd, "-e2") == 0) ||
-        (strcmp(cmd, "-e3") == 0) || (strcmp(cmd, "-e4") == 0) ||
-        (strcmp(cmd, "-e5") == 0) || (strcmp(cmd, "-e6") == 0) ||
-        (strcmp(cmd, "-e7") == 0) || (strcmp(cmd, "-e8") == 0) ||
-        (strcmp(cmd, "--encode") == 0) || (strcmp(cmd, "--encode0") == 0) ||
-        (strcmp(cmd, "--encode1") == 0) || (strcmp(cmd, "--encode2") == 0) ||
-        (strcmp(cmd, "--encode3") == 0) || (strcmp(cmd, "--encode4") == 0) ||
-        (strcmp(cmd, "--encode5") == 0) || (strcmp(cmd, "--encode5") == 0) ||
-        (strcmp(cmd, "--encode7") == 0) || (strcmp(cmd, "--encode8") == 0)) {
+             (strcmp(cmd, "-e1") == 0) || (strcmp(cmd, "-e2") == 0) ||
+             (strcmp(cmd, "-e3") == 0) || (strcmp(cmd, "-e4") == 0) ||
+             (strcmp(cmd, "-e5") == 0) || (strcmp(cmd, "-e6") == 0) ||
+             (strcmp(cmd, "-e7") == 0) || (strcmp(cmd, "-e8") == 0) ||
+             (strcmp(cmd, "--encode") == 0) || (strcmp(cmd, "--encode0") == 0) ||
+             (strcmp(cmd, "--encode1") == 0) || (strcmp(cmd, "--encode2") == 0) ||
+             (strcmp(cmd, "--encode3") == 0) || (strcmp(cmd, "--encode4") == 0) ||
+             (strcmp(cmd, "--encode5") == 0) || (strcmp(cmd, "--encode5") == 0) ||
+             (strcmp(cmd, "--encode7") == 0) || (strcmp(cmd, "--encode8") == 0)) {
 
         // ------------------------------------------------------------------
         // Encode: [-e[0-8] | --encode[0-8]] <lat:-90..90> <lon:-180..180> [territory]>
@@ -519,25 +507,25 @@ int main(const int argc, const char** argv)
 
         // Get territory context.
         int context = 0;
-        const char* defaultTerritory = "AAA";
+        const char *defaultTerritory = "AAA";
         if (argc == 5) {
             context = convertTerritoryIsoNameToCode(argv[4], 0);
             defaultTerritory = argv[4];
         }
 
         // Encode the lat/lon to a set of Mapcodes.
-        char* results[2 * MAX_NR_OF_MAPCODE_RESULTS];
+        char *results[2 * MAX_NR_OF_MAPCODE_RESULTS];
         const int nrResults = encodeLatLonToMapcodes_Deprecated(results, lat, lon, context, extraDigits);
         if (nrResults <= 0) {
             fprintf(stderr, "error: cannot encode lat=%.12g, lon=%.12g (default territory=%s)\n",
-                lat, lon, defaultTerritory);
+                    lat, lon, defaultTerritory);
             return NORMAL_ERROR;
         }
 
         // Output the Mapcode.
         for (int i = 0; i < nrResults; ++i) {
-            const char* foundMapcode = results[(i * 2)];
-            const char* foundTerritory = results[(i * 2) + 1];
+            const char *foundMapcode = results[(i * 2)];
+            const char *foundTerritory = results[(i * 2) + 1];
             printf("%s %s\n", foundTerritory, foundMapcode);
 
             // Self-checking code to see if decoder produces the lat/lon for all of these Mapcodes.
@@ -547,7 +535,7 @@ int main(const int argc, const char** argv)
         }
     }
     else if ((strcmp(cmd, "-b") == 0) || (strcmp(cmd, "-bXYZ") == 0) ||
-        (strcmp(cmd, "--boundaries") == 0) || (strcmp(cmd, "--boundariesXYZ") == 0)) {
+             (strcmp(cmd, "--boundaries") == 0) || (strcmp(cmd, "--boundariesXYZ") == 0)) {
 
         // ------------------------------------------------------------------
         // Generate a test set based on the Mapcode boundaries.
@@ -559,7 +547,7 @@ int main(const int argc, const char** argv)
         }
         if (argc == 3) {
             extraDigits = atoi(argv[2]);
-            if ((extraDigits < 0) || (extraDigits> 8)) {
+            if ((extraDigits < 0) || (extraDigits > 8)) {
                 fprintf(stderr, "error: parameter extraDigits must be in [0..8]\n\n");
                 usage(appName);
                 return NORMAL_ERROR;
@@ -576,15 +564,15 @@ int main(const int argc, const char** argv)
             double lat;
             double lon;
 
-            const mminforec *mm = boundaries(i);            
+            const mminforec *mm = boundaries(i);
             minLon = ((double) mm->minx) / 1.0E6;
             maxLon = ((double) mm->maxx) / 1.0E6;
             minLat = ((double) mm->miny) / 1.0E6;
             maxLat = ((double) mm->maxy) / 1.0E6;
 
             // Try center.
-            lat = (maxLat - minLat ) / 2.0;
-            lon = (maxLon - minLon ) / 2.0;
+            lat = (maxLat - minLat) / 2.0;
+            lon = (maxLon - minLon) / 2.0;
             generateAndOutputMapcodes(lat, lon, 0, extraDigits, useXYZ);
 
             // Try corners.
@@ -613,9 +601,9 @@ int main(const int argc, const char** argv)
         outputStatistics();
     }
     else if ((strcmp(cmd, "-g") == 0) || (strcmp(cmd, "-gXYZ") == 0) ||
-        (strcmp(cmd, "--grid") == 0) || (strcmp(cmd, "--gridXYZ") == 0) ||
-        (strcmp(cmd, "-r") == 0) || (strcmp(cmd, "-rXYZ") == 0) ||
-        (strcmp(cmd, "--random") == 0) || (strcmp(cmd, "--randomXYZ") == 0)) {
+             (strcmp(cmd, "--grid") == 0) || (strcmp(cmd, "--gridXYZ") == 0) ||
+             (strcmp(cmd, "-r") == 0) || (strcmp(cmd, "-rXYZ") == 0) ||
+             (strcmp(cmd, "--random") == 0) || (strcmp(cmd, "--randomXYZ") == 0)) {
 
         // ------------------------------------------------------------------
         // Generate grid test set:    [-g | --grid]   <nrOfPoints> [<extradigits>]
@@ -644,7 +632,7 @@ int main(const int argc, const char** argv)
         if (random) {
             if (argc == 5) {
                 const int seed = atoi(argv[4]);
-                srand(seed);
+                srand((unsigned int) seed);
             }
             else {
                 srand((unsigned int) time(0));
